fix(5th-review): apply correctness and robustness fixes across all modules
Vali-Time:
- EndOf(Day/Week/Month/Year/Quarter) unified to TicksPerDay-1 precision
- EndOfWeek: add DateTime.MaxValue overflow guard
- StartOfWeek: add DateTime.MinValue underflow guard
- PreviousQuarterStart / NextQuarterStart: boundary guards for year 1/9999
- ParseTime: add month (mo) and year (yr) token support to round-trip FormatTime
- Constants: SecondsInMinute/Hour changed from int to decimal

Vali-Calendar:
- WorkdaysBetween: scan limit fixed from MaxWorkdayScan*2 to MaxWorkdayScan (10 years)

Vali-CountDown:
- Breakdown: add TimeUnit.Days key; hours now 0-23 component (not total hours)

Vali-Range:
- ToTimeSpan: clamp decimal product before long cast to prevent overflow on large amounts

Vali-Age:
- Exact(): fix off-by-1 for month-length boundary dates (e.g. Jan31 → Feb28)
- NextBirthday(): guard against year 9999 overflow
- Relative(): update XML doc to reflect ~30.4-day threshold

Vali-Holiday:
- BaseHolidayProvider: GetHolidays now delegates to ConcurrentDictionary cache
- IrelandHolidayProvider: St. Brigid's Day — fix condition Monday→Friday
- IValiHoliday: Register() returns IValiHoliday instead of concrete ValiHoliday
- Mexico: 3 floating-Monday holidays moved to GetMovableHolidays
- Argentina: 2 floating-Monday holidays moved to GetMovableHolidays
- Colombia: 7 Ley Emiliani holidays moved to GetMovableHolidays
- Honduras: hn_morazan_birthday moved to 3rd Monday of October

Vali-Schedule:
- Weekly recurrence: interval check anchored per day-of-week to eliminate drift
- Build(): validate custom predicate is set when RecurrenceType.Custom

Vali-TimeZone:
- GetOffset/OffsetDiff: use UTC→local conversion before GetUtcOffset for DST accuracy
- ResolveZone: Windows StandardName fallback restricted to Windows platform only

Vali-Duration:
- Add FromMonths() and FromYears() factory methods

Vali-Tempo:
- AddValiTempo() now registers AddValiDuration()

All 408 tests pass.

+++ Vali-Time/Utils/Constants.cs
@@ -9,7 +9,7 @@ public  static class Constants
    /// <summary>
     /// Represents the number of seconds in a minute (60). Used for converting between seconds and minutes.
     /// </summary>
-    public const int SecondsInMinute = 60;
+    public const decimal SecondsInMinute = 60m;
 
     /// <summary>
     /// Represents the number of milliseconds in a second (1000). Used for converting between milliseconds and seconds.
@@ -19,7 +19,7 @@ public  static class Constants
     /// <summary>
     /// Represents the number of seconds in an hour (3600). Used for converting between seconds and hours.
     /// </summary>
-    public const int SecondsInHour = 3600;
+    public const decimal SecondsInHour = 3600m;
 
     /// <summary>
     /// Represents the string prefix for milliseconds ("ms"). Used in time formatting.
